% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierReport.R
\name{outlierReport}
\alias{outlierReport}
\title{Examine outliers in a data set using IRQ method}
\usage{
outlierReport(
  dat,
  var_vec = NULL,
  labels = NULL,
  file_path = NULL,
  output_excel = TRUE
)
}
\arguments{
\item{dat}{a data.frame}

\item{var_vec}{NULL or a character vector of colnames to be included}

\item{labels}{NULL or a string that names a colname that holds row-wise ids.}

\item{file_path}{NULL or a string that names the output file.}

\item{output_excel}{a logical.}
}
\value{
an .xlsx file or a list
}
\description{
`outlierReport` returns either an Excel report or a list which
contain information about any mild and severe outliers found in
the data. Classification based on being either 1.5x or 3.0x
outside the IRQ range.
}
\details{
If 'var_vec is NULL, the output will include all of the numeric
variables in the 'dat'. This might be overkill for your application.

If 'label' is provided, the output table will have columns for
'Mild_Outliers' and 'Severe_Outliers' with whatever is contained
in 'label'. This provides a way to track if the same row
is deemed an outlier on multiple numeric columns.

'file_path' must end in '.xlsx'. It's possible to point the file
at an existing sub-directory of the working directory by using something
like './reports/Outlier_Report_v1.xlsx'.

If 'output_excel' is TRUE, the function will try to write an Excel
file report. Otherwise wit will return a list. The list has the same
data from that would be printed in the Excel file, and as well as
a list of the outlier values of each variable.
}
\examples{

 # return a list
 res <- outlierReport(dat = airquality, output_excel = FALSE)

 # return an Excel file
 outlierReport(dat = airquality, file_path = "OutilerTable.xlsx" ,output_excel = TRUE)

}
