% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplotXYZ.R
\name{scatterplotXYZ}
\alias{scatterplotXYZ}
\title{Make a pretty scatter plot}
\usage{
scatterplotXYZ(
  dat,
  x,
  y,
  z = NULL,
  group = NULL,
  point_label = NULL,
  label_size = 3.5,
  min_segment_length = Inf,
  max_overlaps = Inf,
  point_size = 2,
  title = NULL,
  filename = NULL,
  width = 15,
  height = 15,
  units = "cm",
  dpi = 200
)
}
\arguments{
\item{dat}{a data.frame}

\item{x}{a string that names the column in dat to be displayed on the x-axis}

\item{y}{a string that names the column in dat to be displayed on the y-axis}

\item{z}{NULL or a string that names the column in dat to be represented by
the size of the points.}

\item{group}{NULL or a string that names the column in dat that will give colors
to subgroups of point}

\item{point_label}{NULL or a string that names the column in dat that will provide
text labels for the points}

\item{label_size}{a numeric value that controls the size of the labels}

\item{min_segment_length}{a positive numeric value or Inf which controls the
length of segments between the point and label. If 0, then no segments are shown.}

\item{max_overlaps}{an integer from zero to Inf. Larger values increase the
number of text label overlaps that are permitted.}

\item{point_size}{a positive integer that controls the size of the points when 'z'
is NULL}

\item{title}{a string. NULL or a string. If the string is "xy" the plot's title
will use the for of 'x vs. y'. Else, the string will be the title.}

\item{filename}{NULL or a string that names the image file and optionally the
path to the file. The file name must include an image file suffix such as ".jpg"
".png", ".tiff", or ".pdf".}

\item{width, height}{Plot size in units expressed by the units argument. If not
supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the width and height arguments
are expressed: "in", "cm", "mm" or "px".}

\item{dpi}{Plot resolution}
}
\value{
a ggplot2 and optionally a image file
}
\description{
`scatterplotXYZ` returns a ggplot2-based scatter plot in the console,
and optionally as an image file.
}
\details{
The base scatter plot uses x and y values with dark grey points. It can be further
customized to include a z-dimension coded by the size of the points. Colors
can be displayed for a grouping variable. The points can have text labels
with or with segments that connect each label with its point. The plot is
returned as an object and can also be saved as an image file.

When the width and height parameters are set to NA, the output file will be
sized in relation to the plot window. This approach can be helpful for visualizing
the proportions before printing an image file.
}
\examples{

 # z' is NULL and title is "xy"
 scatterplotXYZ(dat = iris, x = "Sepal.Length", y = "Sepal.Width", z = NULL,
                group = "Species", title = "xy")

 # "z" is used to control point size and the title is customized
 scatterplotXYZ(dat = iris, x = "Sepal.Length", y = "Sepal.Width", z = "Petal.Width",
                group = "Species", title = "Iris Scatter Plot")
}
